+++ patterns/WahidWaistcoat/WahidWaistcoat.php
@@ -12,22 +12,91 @@ namespace Freesewing\Patterns;
 class WahidWaistcoat extends JoostBodyBlock
 {
     /**
-     * Add parts in config file to pattern
-     *
-     * I override the pattern's class loadParts() here
-     * because I want to not render the blocks that we're
-     * getting from the parent.
+     * Fix collar ease to 1.5cm
      */
-    public function loadParts()
+    const COLLAR_EASE = 15;
+
+    /**
+     * Fix sleevecap ease to 1.5cm
+     */
+    const SLEEVECAP_EASE = 15;
+
+    /**
+     * Fix pocketWidht to 10cm
+     */
+    const POCKET_WIDTH = 100;
+
+    /**
+     * Fix pocketHeight to 1.5cm
+     */
+    const POCKET_HEIGHT = 15;
+
+    /**
+     * Fix pocketAngle to 5 degrees
+     */
+    const POCKET_ANGLE = 5;
+
+    /**
+     * Fix backNeckCutout to 2.5cm
+     */
+    const BACK_NECK_CUTOUT = 25;
+
+    /**
+     * Sets up options and values for our draft
+     *
+     * By branching this out of the sample/draft methods, we can
+     * set a bunch of options and values the influence the draft
+     * without having to touch the sample/draft methods
+     * When extending this pattern so we can just implement the
+     * initialize() method and re-use the other methods.
+     *
+     * Good to know: 
+     * Options are typically provided by the user, but sometimes they are fixed
+     * Values are calculated for re-use later
+     *
+     * @param \Freesewing\Model $model The model to sample for
+     *
+     * @return void
+     */
+    public function initialize($model)
     {
-        foreach ($this->config['parts'] as $part => $title) {
-            $this->addPart($part);
-            $this->parts[$part]->setTitle($title);
+        // Options that are fixed
+        $this->setOption('collarEase', self::COLLAR_EASE);
+        $this->setOption('sleevecapEase', self::SLEEVECAP_EASE);
+        $this->setOption('backNeckCutout', self::BACK_NECK_CUTOUT);
+
+        // Depth of the armhole
+        $this->setValue('armholeDepth', 290 - self::BACK_NECK_CUTOUT + ($model->m('shoulderSlope') / 2 - 27.5) + ($model->m('bicepsCircumference') / 10));
+        
+        // Collar widht and depth
+        $this->setValue('collarWidth', ($model->getMeasurement('neckCircumference') / self::PI) / 2 + 5);
+        $this->setValue('collarDepth', ($model->getMeasurement('neckCircumference') + $this->getOption('collarEase')) / 5 - 8);
+        
+        // Cut front armhole a bit deeper
+        $this->setValue('frontArmholeExtra', 5); 
+
+        // Overlap at front
+        $this->setValue('frontOverlap', 10); 
+
+        // Some helper vars
+        $this->setValue('chest', $model->m('chestCircumference') + $this->o('chestEase'));
+        $this->setValue('waist', $model->m('naturalWaist') + $this->o('waistEase'));
+        $this->setValue('hips',  $model->m('hipsCircumference')  + $this->o('hipsEase'));
+
+        $waist_re = $this->v('chest') - $this->v('waist');
+        $hips_re = $this->v('chest') - $this->v('hips');
+        if ($hips_re <= 0) {
+            $hips_re = 0;
         }
-        $this->parts['frontBlock']->setRender(false);
-        $this->parts['backBlock']->setRender(false);
+        if ($waist_re <= 0) {
+            $waist_re = 0;
+        }
+        
+        $this->setValue('waistReduction', $waist_re);
+        $this->setValue('hipsReduction', $hips_re);
+        $this->setValue('scyeDart', (5 + $waist_re/10));
     }
-    
+
     /**
      * Generates a draft of the pattern
      *
@@ -41,10 +110,7 @@ class WahidWaistcoat extends JoostBodyBlock
      */
     public function draft($model)
     {
-        $this->buildCore($model);
-        $this->draftFrontPartA($model);
-        $this->draftBack($model);
-        $this->draftFrontPartB($model);
+        $this->sample($model);
 
     }
 
@@ -62,69 +128,20 @@ class WahidWaistcoat extends JoostBodyBlock
      */
     public function sample($model)
     {
-        $this->buildCore($model);
-
-    }
-
-    /**
-     * Drafts the blocks this is based on
-     *
-     * @param \Freesewing\Model $model The model to draft for
-     *
-     * @return void
-     */
-    public function buildCore($model)
-    {
-        $this->setDefaults($model);
+        $this->initialize($model);
+        
         $this->draftBackBlock($model);
         $this->draftFrontBlock($model);
-        $this->draftWaistcoatFrontBlock($model);
-    }
-
-    /**
-     * Presets some defaults
-     *
-     * I'm setting these as options. This way, we could always make them
-     * options later, and the pattern will not require any changes
-     * apart from in this method
-     *
-     * @param \Freesewing\Model $model The model to draft for
-     *
-     * @return void
-     */
-    public function setDefaults($model)
-    {
-        // These could in principle be options
-        $this->setOption('pocketWidth', 90);
-        $this->setOption('pocketHeight', 15);
-        $this->setOption('pocketAngle', -5);
-        $this->setOption('armholeDepth', 290 + ($model->m('shoulderSlope')/2-27.5) + ($model->m('bicepsCircumference')/10));
-
-        // Some helper vars
-        $chest = $model->m('chestCircumference') + $this->o('chestEase');
-        $waist = $model->m('waistCircumference') + $this->o('waistEase');
-        $hips  = $model->m('hipsCircumference')  + $this->o('hipsEase');
-        $waist_re = $chest - $waist;
-        $hips_re = $chest - $hips;
-        if ($hips_re <= 0) {
-            $hips_re = 0;
-        }
-        if ($waist_re <= 0) {
-            $waist_re = 0;
-        }
-        $this->waistReduction = $waist_re;
-        $this->hipsReduction = $hips_re;
-        $this->scyeDart =  5+$waist_re/10;
         
-        // These is irrelevant, but needed for JoostBodyBlock
-        $this->setOption('collarEase', 15);
+        $this->draftWaistcoatFrontBlock($model);
+        
+        $this->draftFront($model);
+        $this->draftBack($model);
 
-        // JoostBodyBlock stores these in $this->help, so we need this
-        $this->help = array();
-        $this->help['armholeDepth'] = $this->o('armholeDepth');
-        $this->help['sleevecapShapeFactor'] = 1;
-        $this->help['collarShapeFactor'] = 1;
+        $this->parts['frontBlock']->setRender(false);
+        $this->parts['backBlock']->setRender(false);
     }
+
     /**
      * Drafts the waistcoatFrontBlock
      *
@@ -146,23 +163,54 @@ class WahidWaistcoat extends JoostBodyBlock
     public function draftWaistcoatFrontBlock($model)
     {
         $this->clonePoints('frontBlock', 'waistcoatFrontBlock');
+        /** @var \Freesewing\Part $p */
         $p = $this->parts['waistcoatFrontBlock'];
 
         // Neck cutout
-        $overlap = 10; // 1cm passed center front
-        $p->newPoint(  300, $p->x(1)-$overlap, $p->y(5) - 80 + $this->o('frontDrop'), 'Neck cutout base point');
-        if ($this->o('frontStyle') == 1) {
+        $p->newPoint(  300, $p->x(1)-$this->v('frontOverlap'), $p->y(5) - 80 + $this->o('frontDrop'), 'Neck cutout base point');
+        if ($this->o('frontStyle') == 2) {
             $p->newPoint(  301, $p->x(8), $p->y(300), 'Neck cutout control point');
         } else {
-            $p->newPoint(  301, $p->x(8), $p->y(9), 'Neck cutout control point');
+            $p->newPoint(  301, $p->x(8)+20, $p->y(9), 'Neck cutout control point');
         }
-  
-        //Hem
-        $p->newPoint(  302, $p->x(4)-$overlap, $p->y(4)+$this->o('lengthBonus'), 'Bottom edge');
 
-        // Dart units
-        $w8th = $this->waistReduction/8;
-        $h8th = $this->hipsReduction/8;
+        // Front inset
+        $p->addPoint( 10, $p->shift(10,180,$this->o('frontInset')));  
+        $p->addPoint( 17, $p->shift(17,180,$this->o('frontInset')));  
+        $p->addPoint( 18, $p->shift(18,180,$this->o('frontInset')));  
+        $p->addPoint( 14, $p->shift(14,180,$this->o('frontInset')/2));  
+        $p->addPoint( 15, $p->shift(15,180,$this->o('frontInset')/2));  
+        $p->addPoint( 16, $p->shift(16,180,$this->o('frontInset')/2));  
+
+        // Shoulder inset
+        $p->addPoint( 12, $p->shiftTowards(12,8,$this->o('shoulderInset')));  
+        //$p->addPoint( 19, $p->shiftTowards(12,8,10));  
+        //$p->addPoint( 19, $p->rotate(19,12,90));  
+
+        // Neck inset
+        $p->addPoint( 8, $p->shiftTowards(8,12,$this->o('neckInset')));  
+        $p->addPoint( 20, $p->shiftTowards(8,12,20));  
+        $p->addPoint( 20, $p->rotate(20,8,-90));  
+
+        
+        
+        //Hem
+        $p->newPoint(  302, $p->x(4)-$this->v('frontOverlap'), $p->y(4)+$this->o('lengthBonus'), 'Bottom edge');
+
+        // Waist reduction
+        $w8th = $this->v('waistReduction')/8;
+        // Hips reduction
+        if($this->v('hipsReduction') < 0) { 
+            // Ease is less than chest
+            // To prevent dart from overlapping,
+            // move excess ease to other seams
+            $h8th = $this->v('hipsReduction')/6;
+            $hDart = 0;
+        } else {
+            // More ease than chest, so divide easy evenly
+            $h8th = $this->v('hipsReduction')/8;
+            $hDart = $h8th;
+        }
       
         // Front dart
         $p->newPoint(  900 , $p->x(5)*0.5,$p->y(3), 'Dart center');
@@ -174,7 +222,7 @@ class WahidWaistcoat extends JoostBodyBlock
         $p->addPoint(  906 , $p->flipX(905,$p->x(900)));
         $p->addPoint(  907 , $p->shift(900,90, $p->deltaY(5,3)));
         $p->addPoint(  908 , $p->shift(900,-90, $p->deltaY(3,4)));
-        $p->addPoint(  909 , $p->shift(908,0, $h8th/2));
+        $p->addPoint(  909 , $p->shift(908,0, $hDart/2));
         $p->addPoint(  910 , $p->flipX(909,$p->x(900)));
         $p->newPoint(  911 , $p->x(909), $p->y(302));
         $p->newPoint(  912 , $p->x(910), $p->y(302));
@@ -209,7 +257,7 @@ class WahidWaistcoat extends JoostBodyBlock
      *
      * @return void
      */
-    public function draftFrontPartA($model)
+    public function draftFront($model)
     {
         $this->clonePoints('waistcoatFrontBlock', 'front');
         $p = $this->parts['front'];
@@ -261,18 +309,21 @@ class WahidWaistcoat extends JoostBodyBlock
         $p->newSnippet("buttonBottom", 'button', 5050);
   
         // Pockets
-        $pw = $this->o('pocketWidth');
-        $ph = $this->o('pocketHeight');
-        $pa = $this->o('pocketAngle'); 
-        $p->newPoint(7000, $p->x(900),$p->y(900)+$p->deltaY(900,4001)*0.6-$ph/2);
-        $p->curveCrossesY(901,905,913,909,$p->y(7000), 700); // Creates point 7001
-        $p->addPoint(7002, $p->flipX(7001,$p->x(7000)));
-        $p->addPoint(7003, $p->shift(7000,-$pa,$pw/2));
-        $p->addPoint(7004, $p->shift(7003,-$pa-90,$ph));
-        $p->addPoint(7005, $p->shift(7000,180-$pa,$pw/2));
-        $p->addPoint(7006, $p->shift(7005,-$pa-90,$ph));
-        $p->curveCrossesY(901,905,913,909,$p->y(7000)+$ph, 701); // Creates point 7011
-        $p->addPoint(7012, $p->flipX(7011,$p->x(7000)));
+        $pw = self::POCKET_WIDTH;
+        $ph = self::POCKET_HEIGHT;
+        $pa = self::POCKET_ANGLE; 
+        $p->newPoint(7000, $p->x(900),$p->y(900)+$p->deltaY(900,4001)*0.6-$ph/2); // Center dart, top
+        $p->curveCrossesY(901,905,913,909,$p->y(7000), 700); // Creates point 7001, Right dart side, top
+        $p->addPoint(7002, $p->flipX(7001,$p->x(7000))); // Left dart side, top
+        $p->curveCrossesY(901,905,913,909,$p->y(7000)+$ph, '.help'); // Approx. right dart side, bottom
+        $p->addPoint(7003, $p->shiftTowards(7001,'.help1',$ph)); // Exact right dart side, bottom. Taking dart angle into account
+        $p->addPoint(7004, $p->flipX(7003,$p->x(7000))); // Left dart side, bottom
+
+        $p->addPoint(7005, $p->shift(7001, $p->angle(7001,7003)-90+self::POCKET_ANGLE, $pw/2));
+        $p->addPoint(7006, $p->shift(7005, $p->angle(7001,7003)-180+self::POCKET_ANGLE, $ph));
+
+        $p->addPoint(7007, $p->shift(7002, $p->angle(7002,7004)-90+self::POCKET_ANGLE, $pw/-2));
+        $p->addPoint(7008, $p->shift(7007, $p->angle(7002,7004)-180+self::POCKET_ANGLE, $ph));
 
         // Make Front shoulder 1cm more sloped
         $p->addPoint('shoulderFront', $p->shiftAlong(12,19,17,10,10));
@@ -284,7 +335,7 @@ class WahidWaistcoat extends JoostBodyBlock
         $p->addPoint(2002, $p->shift(2001,$p->angle(907,2000)-90,$this->o('frontScyeDart')));
         $angle = $p->angle(907,2000) - $p->angle(907,2002);
         $this->msg("angle is $angle");
-        $torotate = array(15,14,16,13,5,2907,2904,2902,2906,2914,2910,2912,911,909,913,905,901,903,7001,7003,7004,7011);
+        $torotate = array(15,14,16,13,5,2907,2904,2902,2906,2914,2910,2912,911,909,913,905,901,903,7001,7003,7005,7006);
         if($this->o('hemStyle')==1) { // Classic hem
             $torotate[] = 4008;
             $torotate[] = 4009;
@@ -295,12 +346,20 @@ class WahidWaistcoat extends JoostBodyBlock
         
         
         // Paths
-        $pocket = 'M 7001 L 7003 7004 7011 M 7012 L 7006 7005 7002';
+        $pocket = 'M 7001 L 7005 7006 7003 M 7004 L 7008 7007 7002';
         $p->newPath('pocket', $pocket, ['class' => 'helpline']);
-        $path = 'M 300 L 302 L 4 L 6 L 5 C 13 16 14 C 15 18 10 C 17 19 12 L 8 C 20 301 300 z M 4001 L 4002';
-        $p->newPath('help', $path, ['class' => 'seam-allowance']);
-        $outline = "M 300 L 4001 $hem L 2910 C 2914 2906 2902 C 2904 2907 5 C 13 16 14 C 15 18 10 C 17 shoulderFrontCp shoulderFront L 8";
+        $outline = "M 300 L 4001 $hem L 2910 C 2914 2906 2902 C 2904 2907 5 C 13 16 14 C 15 18 10 C 17 shoulderFrontCp shoulderFront L 8 C 20 301 300 z";
         $p->newPath('seamline', $outline);
+
+        // Grid anchor
+        $p->clonePoint(302, 'gridAnchor');
+
+        // Mark path for sample service
+        $p->paths['pocket']->setSample(true);
+        $p->paths['seamline']->setSample(true);
+        
+        // Store shoulder seam length
+        $this->setValue('frontShoulderSeamLength', $p->distance(8,'shoulderFront'));
 
     }
     
@@ -317,51 +376,61 @@ class WahidWaistcoat extends JoostBodyBlock
         $p = $this->parts['back'];
         $wfb = $this->parts['waistcoatFrontBlock'];
 
+        // Back inset
+        $p->addPoint( 10, $p->shift(10,180,$this->o('backInset')));  
+        $p->addPoint( 17, $p->shift(17,180,$this->o('backInset')));  
+        $p->addPoint( 18, $p->shift(18,180,$this->o('backInset')));  
+        $p->addPoint( 14, $p->shift(14,180,$this->o('backInset')/2));  
+        $p->addPoint( 15, $p->shift(15,180,$this->o('backInset')/2));  
+        $p->addPoint( 16, $p->shift(16,180,$this->o('backInset')/2));  
+
+
+        // Neck inset
+        $p->addPoint( 8, $p->shiftTowards(8,12,$this->o('neckInset')));  
+        $p->addPoint( 20, $p->shiftTowards(8,12,20));  
+        $p->addPoint( 20, $p->rotate(20,8,-90));  
+        
         // Clone dart points from waistcoatFrontBlock
         $points = array(912,910,914,906,902,904,907,903,901,905,913,909,911,2910,2914,2906,2902,2904,2907);
         foreach($points as $i) $p->newPoint($i, $wfb->x($i), $wfb->y($i));
 
         // Make back shoulder 1cm more sloped
-        $p->addPoint( 12, $p->shift(12,$p->angle(8,12)+90,10));
+        $p->addPoint( 'shoulderBack', $p->shift(12,$p->angle(8,12)+90,10));
+
+        // Shoulder inset
+        $p->addPoint( 'shoulderBack', $p->shiftTowards('shoulderBack',8,$this->o('shoulderInset')));  
         
-        // Bring 1cm from shoulder to front
-        $p->addPoint( 'backShoulder', $p->shift(12,$p->angle(8,12)+90,10));
-        $p->addPoint( 'backShoulderCp', $p->shift(12,$p->angle(8,12)+90,20));
-        $p->addPoint( '.help', $p->shift(8,$p->angle(8,12)+90,10));
-        $p->addPoint( '.help', $p->shift('.help',$p->angle(8,12),10));
-        $p->curveCrossesLine(8,20,1,1,'backShoulder','.help', 'backNeck');
-        $points = $p->splitCurve(8,20,1,1,'backNeck1');
-        $p->addPoint('backNeck2', $points[1]);
-        $p->addPoint('backNeck3', $points[2]);
-        $p->addPoint('backNeck4', $points[6]);
-        $p->addPoint('backNeck5', $points[7]);
+        // Center back dart
+        $p->addPoint(1, $p->shift(1,0,$this->o('centerBackDart')));
+        $p->newPoint('1cp', $p->x(2), $p->y(10));
+
+        // Make shoulder seam seam length
+        $p->addPoint('shoulderBack', $p->shiftTowards(8,'shoulderBack',$this->v('frontShoulderSeamLength')));
+        $p->addPoint( 19, $p->shiftTowards('shoulderBack',8,10));  
+        $p->addPoint( 19, $p->rotate(19,'shoulderBack',90));  
+
+        // Back scye dart
+        $p->addPoint('.help1', $p->shift(10, $p->angle(907,10)-90, $this->o('backScyeDart')/2)); // Half of the dart
+        $angle = 2*($p->angle(907,'.help1') - $p->angle(907,10)); // This is the dart angle
+        $toRotate = [18,10,19,17,'shoulderBack',8,20,1]; // Points involved in dart rotation
+        foreach($toRotate as $i) $p->addPoint($i, $p->rotate($i, 907, -1*$angle));
 
 
         // Paths
-        $path = 'M 1 L 2 L 3 L 4 L 6 L 5 C 13 16 14 C 15 18 10 C 17 19 12 L 8 C 20 1 1 z';
-        $p->newPath('seamline', $path);
+        $partA = 'M 1 C 1 1cp 2 L 3 L 4 ';
+        $dart = 'L 910 C 914 906 902 C 904 907 907 C 907 903 901 C 905 913 909 ';
+        $partB = 'L 2910 C 2914 2906 2902 C 2904 2907 5 C 13 16 14 C 15 18 10 C 17 19 shoulderBack L 8 C 20 1 1 z';
+        $withDart = $partA.$dart.$partB;
+        $withoutDart = $partA.$partB;
+        $p->newPath('seamline', $withDart);
+        $p->newPath('sa', $withoutDart);
+        $p->paths['sa']->setRender(false);
 
+        // Grid anchor
+        $p->clonePoint(4, 'gridAnchor');
+        
+        // Mark path for sample service
+        $p->paths['seamline']->setSample(true);
     }
     
-    /** 
-     * Drafts the waistcoat front
-     *
-     * @param \Freesewing\Model $model The model to draft for
-     *
-     * @return void
-     */
-    public function draftFrontPartB($model)
-    {
-        $p = $this->parts['front'];
-        $b = $this->parts['back'];
-
-        // Clone shoulder points from back
-        $points = array('backShoulder', 'backNeck1','backNeck2', 'backNeck3');
-        foreach($points as $i) $p->newPoint($i, $b->x($i), $b->y($i));
-
-        // Put them in the right place
-        $p->addPoint('backShoulder', $p->rotate('backShoulder','shoulderFront',180));
-        $p->addPoint('backShoulderCp', $p->shift('backShoulder',$p->angle(8,'shoulderFront')+90,10));
-
-    }
 }